+++ core/fresa_config.h
@@ -29,7 +29,7 @@ namespace fresa
         //: name
         constexpr str_view virtual name() const { return "fresa"; };
         //: version
-        constexpr std::array<ui8, 3> virtual version() const { return {0, 4, 7}; };
+        constexpr std::array<ui8, 3> virtual version() const { return {0, 4, 8}; };
         //: unit tests to run (comma separated list)
         constexpr str_view virtual run_tests() const { return ""; };
         //: log level (see tools/log.h for the list of levels)