+++ CHANGELOG.md
@@ -5,7 +5,17 @@
 _new implementation of fresa_
 
 #### todo
-- jobs, events, reflection documentation
+- jobs documentation
+- jobs example
+- events documentation
+- reflection documentation
+- reflection example
+- full "game" example
+
+#### [0.4.8] documentation and final touches (_00 ago 22_)
+
+- **fixed** - strong types can now propperly be formatted
+- **added** - reflection format operator
 
 #### [0.4.7] reflection (_17 ago 22_)